requirements.txt
# Google ADK + Gemini
google-adk>=0.1.0
google-generativeai>=0.5.0

# Config & secrets
python-dotenv>=1.0.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0